Added a merge_params and check_access decorator
The merge_params decorator is very simple and takes self._params and
merges it with params. It's main purpose is to explicitly mark the
views in which it is intended that self._params is merged raw.

The check_access decorator factors out the access checking process
and passes args and kwargs to the access checking methods.

Patch by: Sverre Rabbelier

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%401462