Refactored the picker so that it is more generic
The picker now also works in the case the scope_path of the entity
should not be filtered on. Also, it is now easier to add new
functionality due to the fact that there is a seperate Field type
to handle it now.

Patch by: Sverre Rabbelier

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%401846