Implemented simple group sorted using numerical ordering
See role.py for some example usages.
Note that the reason ToS does not appear seperately is due to the
fact that it is defined as a regular field which does not take the
group field from the one defined in the model.

Patch by: Sverre Rabbelier

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%401997