Important roles can now not resign if they are the last in their group.
Fixes Issue 288.

I've decided to do it outside of the logic because if the need arises to kick the last org admin out of a group we need the ability to do so.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%402387