Use ugettext and shorter variable names for inellegible list
We use the abbreviated variable names in other list views as well to
make it doable to work with them (with the meaning of the
abbreviation explained with a comment after the definition). Bring
the inellegible list up to date with this.

Patch by: Sverre Rabbelier

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%402742