A new model for seeding the database that makes it easier to seed many entities.
new_seed_many is a function that seeds in a different way. Instead of using
redirects, it figures out what the high-water mark of seeding is and
proceeds from there.

This is obviously a half-way step.  I've talked about it with Sverre; I think
I've mentioned it to others.  If we like it, I volunteer to convert everything
to this model (I think there will be savings and simplification when we're all
the way).  If we don't like it, I'll back out this code.  But I'll also be sad,
because this makes it a lot easier to add many entities (which I care about
because I'm trying to fix list view for that many entities).

Patch by: Dan Bentley

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%402769