Added the GHOP modules to the callback.
The ENABLE_MODULE has been added to ensure that the current build does not suffer from any weird artefacts that might occur due to the presence of this module. Of course in time it will be removed.

Thanks Madhusudan for getting us this far ^_^. Minor CSS and image patches should follow shortly.