Create context dictionary with empty list before hand for Org home page.
This is necessary because if we are outside the timeline checks i.e before
tasks become visible or after tasks deadlines pass there is no context passed
and hence no list key in the dictionary. So the check for list context variable
in the home template fails. So to avoid that we create a context dictionary
with an empty list as value for the key list.