Adding a page to list all the tasks depending on the role.
For the Program Admin this will list all the tasks except for those in the
Invalid state including the Unapproved and Unpublished tasks for all the
organizations. But for other users it lists only the published tasks but
in any other state except for Invalid state. Org Admins or Mentors who
would like to see the Unapproved and Unpublished tasks for their
organizations should use a separate list which appears under their
respective Organizations menu on the sidebar.