Move button state update for editing during dirty fields check.
Previously it was issued during normal button state updates.
This meant that the button was updated only once another row
was selected, and not when the user pressed Enter,
probably because the dirty fields array update was
done asynchronously and didn't get updated in time for the
button updating function to detect it.