Always surround survey form in a tbody element.
A table/tbody element was automatically inserted presumably by the
browser when the trs and tds from survey form were detected.
If the survey form was empty, then no tbody element was added,
thus preventing survey edit javascript to bind the widget to
the first tbody of the page.

Fixes issue 914.