Refresh lists after a post or post_edit action.
Formerly there was only the "table" option for refreshing
tables (which refreshed only the current list). To reflect
needs, e.g. of the task lists for a GCI organization, this
adds also the "all" options (which refreshes all the lists
in the current page). Furthermore, putting an integer in it
would refresh the specific list.