Move all the boiler plate code required to download a blob to helper.
The current code in Work Submission View to download the blob is mostly
boiler plate code which can be reused by any other view which needs
to download the blob. So this code is now moved to the blob views helper.
So the view just exists to do the right access check catch the blob key
from get request.