GCI student forms download access checks made more robust.
We had hard coded the access checks for downloading to check each form
that GCI student submits. We ran into a problem that, when a new form
is required of the student, we will have to run into access check method
every time which is a pain. Now we instead pass the dictionary
containing the names of the form that student should upload as the keys
and the names of the blob properties which store the forms that students
upload as the values for the keys. We use this dictionary in the access
checks to see if the blob requested for belongs to what student has
submitted.