Convert the MyProjectsComponent to work with the new list code.
This wrecks all other lists in the dashboard but this will help us to figure out where we can improve on the list interface.

Added two new templates to help with rendering of the lists. I'm open for comments on the listS.html since I feel it goes against our current templating philosophy. It works for now though and I needed to get this out so we can play around with the new lists.