Refactor the participatingOrgs logic to fetch last few orgs as well.
We now fetch each org, test if it has the Logo URL set, and if so add
to the list of participating orgs whose logo must be displayed on the
home page.

Before this fix, we always threw away last few orgs if the total number
of orgs returned in the last query was less than the batch size. We
do fetch the remaining orgs now by querying again.