In the additional orgs that are fetched as new query don't add same organizations again.
When the number of organizations which have Logo URL set are less than the
batch size, there is a possibility that same organization gets added again.
This code circumvents that problem.

It is all becoming expensive but we are trading it off since we are doing
all this only once in 30 mins (the cache expiry time).