Just comment the buggy line before we fix it.
I am extremely sorry for commenting the code guys, but it is taking
me time to figure out how to fix this bug. So to keep the production
running I have done this commit. Will fix the bug as soon as I can
make out how to do that.

I apologize again for this commit.