Explicitly teach Python how create tuples, don't let it go according to its whims and fancies.
Effing Python. I so love the language (I understand the significance of the
statement I am making). But as they say too much of love leads to hatred :P.
If there are two things if I hate about Python this is the second in the
list. First one being the way Python handles the authenticated proxies.
(Hope some Python developers are seeing my cry here.)

Python does the implicit conversion of any objects separated by commas to
a tuple containing the same objects. Although this is very useful while
packing an unpacking the tuples, esp swapping like a, b = b, a, Python
does this implicit conversion as it wishes. This caused our production
site to be broken for hours. So making the tupling implicit and uncomment
the code I had previously commented.