Add metaclasses, classes and methods to construct readonly pages for models.
We use the same approach that Django uses to build forms from models in its
ModelForm APIs. So this idea is completely inspired by Django's ModelForm
APIs and the Appengine's monkey patches to support them on Appengine models.

We also use the option names that are consistent with Django's option names
in ModelForm APIs. In addition we also add methods to be consistent with
Melange framework's APIs to render Views.