Check for the datastore write in RequestHandler dispatch method and take the necessary action.
If the request made is a POST request and the datastore writes are disabled
we catch than and redirect the user back the referrer URL with the get parameter
stating that the datastore write capability has been disabled.

In the base template we check if this context is set from the default contexts
and display the appropriate error message on the user screen.

--HG--
extra : rebase_source : c8a838e8384055ea3a21f051fe9f8e0880a7ad55