Fix the bug in the test.
:P I am sorry about this. I had a misconception about the way form errors are
handled. I now realize that when the form has an error, to retain the error
states of fields we cannot redirect to a new GET request since by doing that
we lose the form error states. Instead we send the HTTP OK response for the
same post request with error states set. I have fixed the test to reflect this.