Send the response of the propsal review page when the comment form fails to save.
Absolutely hacky solution. I suggest a solution based on the PJAX concept for loading
comments and its form. But it requires quite some time to implement. This is a
temporary solution until we implement PJAX or find a better solution.

The solution reverses the view name for propsal review page within which comments
and its form are embedded, uses that reversed URL to resolve to the view function,
which is a RequentHandler object in Melange and sneekily embeds the request
object, args and kwargs and calls that view method and return the response.

VERY IMPORTANT POINT TO NOTE: This is NOT unacceptable though. Django's documentation
itself uses this idea to send the response although in a different situation. The
link to documentation is here:
http://docs.djangoproject.com/en/dev/topics/http/urls/#django.core.urlresolvers.ResolverMatch

Web is all about sending an HTTP Response for an HTTP Request right? Who cares how
the responses are built? :P

--HG--
extra : rebase_source : 16832266ba987fa57d9757a26f5f12470a55a763