Use IPython when it is installed.
This was the most painful part for me in the entire Melange development
process. I was not able to use the awesome features of the most powerful
IPython shell like tab completion, code history and other magic commands,
when using remote APIs since the vanilla interpreter doesn't provide
any of these features. Upon this change, if the system has IPython, installed
the interactive scipt uses it instead of the vanilla interpreter. If IPython
is not installed, it reverts back.

Although I have been doing this on almost all the other projects I work on,
I did not have enough to time to embed this code snippet in Melange earlier.
Finally here it is!

And thanks to Django for the snippet. I always get this snippet from Django :)