Do not calculate next deadline for the program that is not active.
It anyways doesn't make sense to check the deadline for the program that
is not active.

However there is one special case. A program that has not started may have
the program as inactive but the org applications are yet to open. So
that case is checked even if the program is active.

--HG--
extra : rebase_source : 71b89ba6d6e6ab528ab4f0a1d295f99506eb5596