Make a copy of assign mentor template outside proposal template directory.
We will have to do this because Assign Mentor action can be performed both
on the proposal and the project pages. So as a first step, I am making a
copy of it outside proposal templates directory. In the subsequent commits,
the one which exists inside the proposal directory will be removed when
refactoring to new action menu.

--HG--
extra : rebase_source : a8b384e50e17a8000835ebdd3ce882e41a5827d1