Define a new RequestHandler for assigning mentor to a project.
This RequestHandler is somewhat similar to the RequestHandler in the proposal
review page. As we factored out the Templates for generating the assign mentor
dropdown list and the UI items, I made a sincere attempt to refactor this code.
But as I progressed through the refactoring I realized that I need to either
make the refactored code way too generic to be used at only 2 places, one
being the proposal review page and the other being project details page or
write it in such a way that only very very few common elements should be
kept in the parent class and derive two separate classes out of it for the
respective pages. In either case, the code becomes less readable and it is
definitely not worth the refactoring I felt, because url patterns, access
checks must all be defined in either cases anyways. So I reverted back all
the refactored code and I am now using 2 different RequestHandlers which
are similar but not really refactorable.

We also register this RequestHandler in the callback.

--HG--
extra : rebase_source : dee0d265015f91e3cd1f01ca7169e1d6ceffb6f0