Make the toggleButton function a self contained object binding to change events.
The toggle buttons are now Javascript object types, which are highly self
contained. They store the state and all initialized variables per object.
Each object created from this will be a toggle button.

We also bind onchange event to perform an AJAXy POST call with the state value
to the defined post url to update the state of the action represented by such
buttons.