Make toggle button template directly accept a URL than the view name.
This is necessary because, since this class is a generic class to generate
toggle buttons. So it is not easy to determine which redirector helper
method to call to populate the args and kwargs, if we supply the view name
to be reversed to the URL. So it is much easier to get the URL where this
class is instantiated since all the args and kwargs necessary to construct
the URL is available in the RequestData object there.