Make the Uniform JS/CSS apply to the only checkbox that appears on the page as is, i.e. make comment private.
This is a very important change and which made me waste 2 full days wondering
what was happening with the new awesome buttons. The reason is that jQuery
Uniform used to apply to all the checboxes on the form and the awesome button
magic works in such a way that we first put an HTML input checkbox tag which
is magically replaced with the new awesome buttons on the Javascript side.
The Uniform Javascript was applying on top of this which made CSS leak from
both Uniform and awesome toggle buttons.

In order to circumvent this problem we very specifically apply Uniform only
to make the comment private checkbox. Now the awesome toggle buttons CSS
and Javascript behave as expected.