Use the melange.map module which uses the Google Maps API v3 in organization home page.
We load the Javascript supplied by Google Maps API v3 here since this cannot
be loaded as a dependency with LABjs dependencies for the reasons explained
in the comment in the code. We also load the CSS necessary for maps.

--HG--
extra : rebase_source : d95697c878e31eef0caf39bed3390632279df8b9