Replace the Survey template to use the jQuery formbuilder plugin.
We remove the HTML, CSS and JS code that was supposed to use the home-brewn
form builder code for surveys. We now implement the HTML, CSS and JS
required to just use the formbuilder plugin API. So we call formbuilder
plugin and initate the form builder code on the HTML place holders
we define.