Store the patch that sends XSRF token also with the form data in jquery formbuilder plugin.
This patch was written for jQuery formbuilder plugin v0.2. If the subsequent
versions of the plugin ever provides an API to add custom data to post, we
an get rid of this patch.

But until then we will have to apply this patch everytime we update jquery
formbuilder plugin to make sure we pass on the XSRF token along with the
formbuilder POST data, we otherwise get an AUTHORIZED REQUEST HTTP Response,
response code 403. But now the XSRF failure should be warned by the
corresponding middleware which does the check for XSRF.