Update the jquery formbuilder plugin to the latest version.
These changes were provided to us by EchoDitto. Many of the changes
here are the update to the plugin itself as they prepare for the
release of the new version. In addition EchoDitto has also applied
some customizations for us.

NOTE: The version of this plugin is still in Beta phase and may
have issues.