Apply the updated CSS that EchoDitto provided.
I directly tried to backport only the changes/diffs to {less} and then
compile CSS from that. But I could not get the desired effects on all
the styles that were supposed to be applied. So to play safe, I have
copy pasted the entire CSS and made some minor changes. I will backport
it to {less} later.