A partial rewrite of the jquery formbuilder plugin.
With this commit I have decided not to store the patches for this plugin
anymore. The original plugin so doesn't suit us that I had to rewrite
some large chunks of code in there. So it stands to be Melange formbuilder
plugin henceforth.

This code itself can become a new jquery plugin. To be nice to the developers
who gave us this plugin, when I have more time, I will send the pull request
for the changes I have made which may be useful to others too on their
github repository. But if they do not want to merge these changes, I will
release this plugin to public myself. A jquery plugin contribution from
Melange developers! :)

*NOTE: Please don't update this plugin with new version and expect the
same behvaior. It has been partially rewritten.*