Remove prefix from survey and survey record URL patterns.
Since we have this URL pattern defined in the GSoC module, it should
have a gsoc_program as prefix. I am yet to see a GSoC program related
survey, including OrgApp surveys which have any other prefix than
gsoc_program. So I think it is a good idea to even completely get
rid of that prefix idea and make our lives easier.

For now, until we completely get rid of prefix from our Survey data
models, the idea is to hard code the value 'gsoc_program' while
creating the entity and not to have it in URL because /gsoc/...
in the URL should itself indicate that it is GSoC related survey.