Modify the survey_record redirector to not have prefix anymore since it is not necessary.
Also change the survey entity passed as argument to this method to
the link_id. This is because when the survey entity is not passed
as argument to the method, the method looks in the request kwargs
for the survey link_id so there will be a datatype mismatch. So to
be consistent pass survey's link_id itself as argument.