Specify that the data we are supplying to jquery formbuilder is JSON serialized.
This is not the feature that we introduced, but is provided in the new
version of the plugin itself. It is very interesting to see how the
data we serialized by the new JSON serializer we have written exactly
matches the required unserialized JSON required to load JSON form.
JSON has implied standards! And that occurs to people intuitively, it
is not needed to teach them! This is the live example for that claim.
The serializer written by some one exactly unserializes the same way
some one else expected. I love JSON!