Change the serialization format to store the built form fields along with their unique IDs.
The current serialization only holds an array of dictionaries where each
dictionary contains the meta data for the form fields that are built. And
the order of the field dictionaries in the array is the order of the form
fields. Now to accommodate unique IDs being assigned to each of this form
field so that it can map to datastore property we have to generate and
store unique IDs to each of the form field built. So we change the
datastructe to store contain an array of two elements where the first
element itself is another array which stores the generated unique ID for
each form field added and the second items is a dictionary where each
key is the unique ID for the field and the corresponding value is the
field dictionary containing the field meta data.

--HG--
extra : rebase_source : b3c4b9125fefd68cac3fb306da4ec717615723ec