Rename the evaluation components to have consistent names.
We change the name of survey which is administered to students to student
evaluation and the name of the survey which is administered to the
organization/mentors to mentor evaluation. We also change the corresponding
URLs to be consistent with this name and make it a bit shorter.

--HG--
extra : rebase_source : 13ec1215eda2fa7d4f8573b9fa620cef34977dbe