Modify the methods that build the form for administering evaluations based on new serialization format.
In this format, we obtain the names for each field from the unique ID
assigned to each field in the form when creating the form structure
instead of generating the names locally from the field label, which may
have ill-effects when people have already started submitting evaluations
and the label needs to change (for example, in cases such as typo fixes).

We also get the fields order from the serialized data itself.