Rewrite the assign mentor portion of select widgets template.
We rewrite these portions to render more than one select widget if there
are multiple mentors assigned to a project. We also make sure that this
combination of multiple select widgets, one for each mentor is stored in
a div of itself so that we can clone and append the select widget into
this div container enclosing all the select widgets to assign mentors.