Add the necessary Javascript to clone the assign mentor select widget.
To keep the ID unique for each select widget added, we take the ID of
the first select widget from which we clone and append a counter which
increments upon each addition and assign it as the ID for the new select
widget. In addition we also apply Uniform CSS and JS to each of these
select widgets.