Change all the occurences of the usage of project evaluation and survey to mentor and student evalutions.
This has been done to have a more consistent naming for evaluation all
across Melange from the URLs to all the usages in evaluations. Only
exceptions that have been left out are urlpatterns module and redirect
helpers because they are more general and can be used across different
types of surveys if anyone wants to create such surveys.