Make use of the mentors property where needed and possible.
Disabled map on the org home page since it needs to now be able to draw multiple lines from one student to the mentors.

Put some question marks around the withdraw view's assumptions that can not be guaranteed afaik.

This change does not include any fixes to lists in the admin, org_home, projects_list and withdraw_projects view.