Use the evaluation row adder function in combination with RawContentResponseBuilder.
We do not have to inherit a new ContentResponseBuilder class for evaluations
if we follow this approach. We can instead use the function which accomplishes
the same. Now the row adder function for evaluations pass eval as the first
positional argument instead of key word argument. Use that in every function
definition instead of evaluation argument from kwargs.