Add a mapper function to convert old survey schema to the new schema.
This function is survey agnoistic, it works on all the survey models
that inherit from the base Survey model.

We need to supply the right entity for conversion for each model separately
by adding a mapreduce entry in the mapreduce.yml for each model. This
limitation exists because our base Survey model is a regular model and
is not a Polymodel. If it were a Polymodel we could have fetched all the
entities that belong to the models inherited from the given Polymodel.