Allow the users who has access to the survey to access it even when the project has failed.
Before this access check is called, we ensure that the deadline has not passed
for submitting the evaluation. If it has passed we redirect the user to the
evaluation summary page. So here we can safely assume that access to editing
survey is not given for failed projects after the deadline has passed.

However, since we use the same access check methods for both edit and show
pages this page lets the user to access the page (show page) past the deadline
even if the project has failed.