Mapreduce to convert the GCITasks from keyname based to ID based.
In the process we recreate all the GCIComment and GCIWorkSubmission entities
because they have GCITask as parents. So we recreate them with the new
parent.

We move the subscribers property from GCITaskSubscription model to the
task model since there is no real need to have a separate model for it.

We also convert the GCIStudentRanking entities which contains the references
to list of tasks to update to the new entities.

NOTE: We copy the link_id from the old GCITask entities to the new one
although we will not use them because for replacing the old GCITask entities
with the new one in the GCIStudentRanking entities we need a way to identify
the new GCITask entities in the datastore that corresponds to the old entities.
Once the conversion is run, we can eliminate Link ID completely.