Build the dictionary of bound fields for the form.
As mentioned in one of the earlier commits, the entire stack of task
creation is highly customized and doesn't follow the patterns generally
seen in other forms in Melange. Melange's ModelForm class generates
BoundFields for each field only when we iterate over the form using the
iterator protocol. However, this is not possible for task creation form
because of the layout constraints. So instead we bind all the fields to
its Bound Field instance during the form creation stage itself.

Attempt to mimic Django's pattern of storing the BoundField instances
as the Form class' attributes with the name of the fields failed because
Django overwrites these attributes with the object of type Django's
BoundField along its path before bailing out the form to the template.
So the bound fields had to be stored in the dictonary.