Fix the backup admin field for Org Apps.
The backupadmin field now raises an error when a user tries to enter himself when creating a new app or when the user tries to enter the id of the person who created the app.

Also the backup_admin_id field will not be present if the cleaning methods raised an exception and popping it from cleaned_data must therefore occur under safe conditions.