Added view for the GCI age check.
It sets age_check to 0 or 1 depending on whether the user indicates he's old enough. If the result is 1 the user is also redirect to the student application page or otherwise the user is shown an error message on this page.

The access check for the student registration page which deals with the cookies will be added soon.