Remove user and history property from GCITask model.
The history property was always a pita to keep up to date. The idea itself was
nice to gather more advanced statistics but the effort to make it work is just
too big.

The user property is not needed anymore since we will have profiles for any
interaction with a task, ergo we can use the student property. The 16 tasks
from GCI2010 which have a user assigned but no student will still have to
render correctly but this should not be a big issue since the data can not
change anymore.