Access checks now give 404 (from 403) when program is not found.
We should actually give a 404 when a program is invisible as well but it currently is clear that the program is there since the menu will properly appear. So an AccessViolation suffices for now.

Suggestion is to build an error page which can work without any program so that we can show that when a program is invisbile and no one will be any wiser.