Very initial implementation of row_buttons added to lists module.
It allows to add either simple redirect button or costum redirect one. From protocol's perspective they do not really differ, as the complete link is set before the row is sent to the browser. From backend perspective. The simple one receives a link as its input, while the costum one receives a function which may generate the link based on the entity.
Also, the idea is that when the function returns None or False, the button should not be added for the entity at all.